feat: specify `baseUrl` and `keyserver`
also, fixed logging in with `CloudAuthBasicUI` still happening when the user authenticates with Better Auth successfully, logs out in Jazz, then tries to log back in to the Better Auth server with incorrect credentials.
note: with this change, if the key shards are lost in the keyserver (eg, it restarts), then you cannot log back into the Better Auth account.